Завершаем работу над проектом. Исправления
