+++ add.php
@@ -4,7 +4,6 @@ require_once('helpers.php');
 require_once('functions/common.php');
 require_once('init.php');
 require_once('data.php');
-require_once('sql_queries.php');
 
 $categories = get_all_categories($link);
