+++ functions/db.php
@@ -24,48 +24,6 @@ function db_fetch_data($link, $sql)
 }
 
 /**
- * Функция получения значения из параметра запроса
- * @param string $lot_id строка с наименованием параметра запроса, получение лота по его id
- * @return string значение параметра запроса
- */
-function get_lot_by_id($lot_id)
-{
-    global $sql_lot;
-    global $link;
-    $sql_lot = sprintf($sql_lot, $lot_id);
-    $result = mysqli_query($link, $sql_lot);
-    if ($result) {
-        if (mysqli_num_rows($result)) {
-            return mysqli_fetch_all($result, MYSQLI_ASSOC)[0];
-        } else {
-            return null;
-        }
-    }
-    return null;
-}
-
-/**
- * Функция получения значения из параметра запроса
- * @param string $email строка с наименованием параметра запроса, поиск по email
- * @return string значение параметра пост-запроса
- */
-function get_user_by_email($email)
-{
-    global $sql_mail;
-    global $link;
-    $sql_mail = sprintf($sql_mail, $email);
-    $result = mysqli_query($link, $sql_mail);
-    if ($result) {
-        if (mysqli_num_rows($result)) {
-            return mysqli_fetch_all($result, MYSQLI_ASSOC)[0];
-        } else {
-            return null;
-        }
-    }
-    return null;
-}
-
-/**
  * @return array|bool|mysqli_result - выполнение запроса на получение всех категорий, обработка запроса
  */
 function get_all_categories($link)
@@ -170,39 +128,6 @@ function get_active_lots_by_category_id($link, $id, $limit, $offset)
 }
 
 /**
- * Функция навигации по страницам
- * @param $url - Конкретный лот в своей конкретной категории
- * @param $cur_page - Конкретная страница
- * @param $pages_count - Количество страниц
- * @return array - Возврат массива значений
- */
-function get_navigation_links($url, $cur_page, $pages_count)
-{
-    return [
-        !($cur_page - 1) ? '#' : $url . '&page=' . ($cur_page - 1),
-        ((int)$cur_page === (int)$pages_count) ? '#' : $url . '&page=' . ($cur_page + 1)
-    ];
-}
-
-/**
- * Функция подсчета и показа лотов на странице, выполнение пагинации
- * @param $lots_count - Подсчет количества лотов
- * @param $cur_page - Конкретная страница
- * @param $page_items - Количество странци
- * @return array - Возврат массива значений
- */
-function get_navigation_data($lots_count, $cur_page, $page_items)
-{
-    $pages_count = ceil($lots_count / $page_items);
-
-    return [
-        $pages_count,
-        ($cur_page - 1) * $page_items,
-        range(1, $pages_count)
-    ];
-}
-
-/**
  * Функция выполнения запроса на получение id всех лотов, у которых дата окончания лота < текущей, а также поле is_closed = 0
  * Обработка данного запроса
  * Получение массива
@@ -219,9 +144,48 @@ function fill_lot_winners($link)
     foreach ($lots as $key => $value) {
         $id = $value['id'];
         $sql_winner_update = "UPDATE rates SET rates.is_winner = 1 WHERE rates.lot_id = '$id' ORDER BY rates.price DESC LIMIT 1";
-        $sql_winner = mysqli_query($link, $sql_winner_update);
+        mysqli_query($link, $sql_winner_update);
     }
 
     $sql_lots_update = "UPDATE lots SET lots.is_closed = 1 WHERE lots.date_of_completion < CURDATE() AND lots.is_closed = 0";
-    $sql_closed = mysqli_query($link, $sql_lots_update);
+    mysqli_query($link, $sql_lots_update);
+}
+
+/**
+ * Функция поиска емейл в базе данных
+ * @param $link - Соединение с базой данных
+ * @param $email - Искомое значение
+ * @return |null - Возврат результата в случае ненахождения
+ */
+function get_user_by_email($link, $email) {
+    $sql = "SELECT * FROM users WHERE email = '%s'";
+    $sql = sprintf($sql, $email);
+
+    $result = mysqli_query($link, $sql);
+
+    if ($result) {
+        if (mysqli_num_rows($result)) {
+            return mysqli_fetch_all($result, MYSQLI_ASSOC)[0];
+        }
+    }
+
+    return null;
+}
+
+/**
+ * Функция нахождения емайл, в случае совпадения по емайл - возврат нулевого результата
+ * @param $link - Соединение с базой данных
+ * @param $email - Искомое значение
+ * @return |null - Возврат результата в случае нахождения
+ */
+function get_user_by_email_result($link, $email)
+{
+    $result = get_user_by_email($link, $email);
+    if ($result) {
+        if (count($result) > 0) {
+            return $result;
+        }
+    }
+
+    return null;
 }
