+++ templates/lot.php
@@ -12,7 +12,7 @@
         <div class="lot-item__right">
                 <div class="lot-item__state">
                     <div
-                        class="lot__timer timer <?php echo (!empty($time_report['hours'] === '00') ? 'timer--finishing' : ''); ?>">
+                        class="lot__timer timer <?php echo (int) $time_report['hours'] === '00' ? 'timer--finishing' : ''; ?>">
                         <?php echo is_array($time_report) ? implode(':', $time_report) : 'error'; ?>
                     </div>
                     <div class="lot-item__cost-state">
